requirements.txt
django
djangorestframework
django-cors-headers
requests
djoser
mysqlclient
pillow
django-froala-editor